+++ hutool-core/src/main/java/cn/hutool/core/io/watch/WatchMonitor.java
@@ -55,10 +55,10 @@ public class WatchMonitor extends Thread implements Closeable, Serializable{
 	public static final WatchEvent.Kind<?> ENTRY_DELETE = StandardWatchEventKinds.ENTRY_DELETE;
 	/** 全部事件 */
 	public static final WatchEvent.Kind<?>[] EVENTS_ALL = {//
-			StandardWatchEventKinds.OVERFLOW,        //事件丢失
-			StandardWatchEventKinds.ENTRY_MODIFY, //修改
-			StandardWatchEventKinds.ENTRY_CREATE,  //创建
-			StandardWatchEventKinds.ENTRY_DELETE   //删除
+			OVERFLOW,        //事件丢失
+			ENTRY_MODIFY, //修改
+			ENTRY_CREATE,  //创建
+			ENTRY_DELETE   //删除
 	};
 	
 	/** 监听路径，必须为目录 */